add useEffect to check for user returning from OAuth sign in, save to db
When PrelaunchSignUpForm is rendered, user may be visiting for the first time or may have just returned from signing in via OAuth. This useEffect checks for whether they are returning from OAuth sign in and, if so, sets flowState to 'processing', invokes createUserDoc, and sets flowState to 'confirmed'. Added setFlowState to dependency array to satisfy eslint rule 'react-hooks/exhaustive-deps'