set custom screen size breakpoint for input text
This is a quick workaround that needs revisiting. Dark mode styling is currently not being applied correctly to screens 390px and below, but input text styling is applied inline. This caused an issue where dark mode styling for input text was being applied without the rest of dark mode styling. Setting a custom breakpoint of 391px for the input text fixes that issue for now.